Fix the same signed-bitwise finding in the new test's own literal
The previous commit ran clang-tidy on VersusFlow.cpp before pushing but not
on the test file it also touched: `(1 << 30) + 1` in the new regression test
is the identical bugprone-signed-bitwise pattern as kMaxBracketSize, just
never checked. Same fix, same reasoning.

Verified against the pinned clang-tidy 23 and clang-format 23, and recompiled
and reran the real test binary (moc + g++ against system Qt): all 29 cases
still pass.

+++ tests/unit/tst_versus_flow.cpp
@@ -273,7 +273,7 @@ void TestVersusFlow::restoreRejectsABracketSizeBeyondWhatRoundIndexingCanShift()
     // pass the length check that follows it, so the bound is checked first:
     // this stays cheap to test and never depends on constructing that array.
     QJsonObject payload;
-    payload.insert(QStringLiteral("bracketSize"), (1 << 30) + 1);
+    payload.insert(QStringLiteral("bracketSize"), static_cast<int>(1U << 30U) + 1);
     payload.insert(QStringLiteral("slots"), QJsonArray{});
 
     domain::VersionedFlowState saved;
